抽出sso-core 修改认证中心无需管理子系统中token 由子系统统一管理 只需子系统验证ssoToken

+++ sso-server/src/main/java/com/sso/server/service/RedisOperatorService.java
@@ -4,9 +4,9 @@ import com.sso.server.entity.TokenSession;
 
 public interface RedisOperatorService {
 	
-	void putUserInfo(String userName, String token, int expireTime);
+	void putUserInfo(String userName, String token, long expireTime);
 	
-	void putTokenInfo(String tokenKey, TokenSession tokenSession, int expireTime);
+	void putTokenInfo(String tokenKey, TokenSession tokenSession, long expireTime);
 	
 	String getUserInfo(String userName);
